issue #88 Calendar DataBinding
